[Refactor] 5감사, 일기 Entity class명 변경, 파일구조 변경
